Payments Module Further Modified

+++ src/app/modules/payment/payment-details/payment-details.component.ts
@@ -3,8 +3,9 @@ import { Router } from '@angular/router';
 import { Payment } from 'src/app/core/model/payment.model';
 import { ConfirmationService, Message, MessageService } from 'primeng/api';
 import { UserService } from 'src/app/core/service/user.service';
-import { ClientService } from 'src/app/core/service/client.service';
+import { VendorService } from 'src/app/core/service/vendor.service';
 import { PaymentService } from 'src/app/core/service/payment.service';
+import { Vendor } from 'src/app/core/model/vendor.model';
 
 const confirmMsg = 'Do you want to delete this payment?';
 
@@ -18,10 +19,12 @@ export class PaymentDetailsComponent implements OnInit {
 	payments: Payment[] = [];
 	tableHeaders: any[];
 	msgs: Message[] = [];
+	vendor: Vendor;
 
 	constructor(
 		private _paymentService: PaymentService,
 		private router: Router,
+		private _vendorService: VendorService,
 		private confirmationService: ConfirmationService,
 		private messageService: MessageService
 	) {}
@@ -41,11 +44,12 @@ export class PaymentDetailsComponent implements OnInit {
 		this.getPayments();
 
 		this.tableHeaders = [
-			{ field: 'clientName', header: 'Client Name' },
+			{ field: 'vendorName', header: 'Vendor Name' },
 			{ field: 'contactPerson', header: 'Contact Person' },
 			{ field: 'phoneNumber', header: 'Phone Number' },
 			{ field: 'amountPaid', header: 'Amount Paid' },
-			{ field: 'PaymentDate', header: 'Payment Date' }
+			{ field: 'PaymentDate', header: 'Payment Date' },
+			{ field: 'PaymentType', header: 'Payment Type' }
 		];
 	}
 
@@ -56,22 +60,42 @@ export class PaymentDetailsComponent implements OnInit {
 	}
 
 	deletePayment(payment: Payment) {
-		this.confirmationService.confirm({
-			message: 'Do you want to delete this payment?',
-			header: 'Delete Confirmation',
-			icon: 'pi pi-info-circle',
-			reject: () => {
-				this.msgs = [ { severity: 'info', summary: 'Rejected', detail: 'You have rejected' } ];
-			},
-			accept: () => {
-				this.msgs = [ { severity: 'info', summary: 'Confirmed', detail: 'Payment Deleted' } ];
-				this._paymentService.deletePayment(payment);
-				this.messageService.add({ severity: 'success', summary: '', detail: 'Payment Deleted' });
-			}
+		this._vendorService.getVendorById(payment.vendorId).subscribe((response) => {
+			response.amountBalance = response.amountBalance + +payment.amountPaid;
+			this.vendor = response;
+			console.log('response.amountBalance', response.amountBalance);
 		});
+		if (payment.billPaymentType == true) {
+			this.messageService.add({ severity: 'info', summary: '', detail: 'Payment Cannot be Deleted' });
+		} else if (payment.manualPaymentType == true) {
+			this.confirmationService.confirm({
+				message: 'Do you want to delete this payment?',
+				header: 'Delete Confirmation',
+				icon: 'pi pi-info-circle',
+				reject: () => {
+					this.msgs = [ { severity: 'info', summary: 'Rejected', detail: 'You have rejected' } ];
+				},
+				accept: () => {
+					this.msgs = [ { severity: 'info', summary: 'Confirmed', detail: 'Payment Deleted' } ];
+					this._paymentService.deletePayment(payment);
+					this._vendorService.updateVendor(this.vendor);
+					this.messageService.add({ severity: 'success', summary: '', detail: 'Payment Deleted' });
+				}
+			});
+		}
 	}
 
 	editPayment(paymentId: string) {
 		this.router.navigate([ 'Payments/New Payment', paymentId ]);
+	}
+
+	getPaymentType(payment): string {
+		if (payment.billPaymentType == true) {
+			return 'Bill Payment';
+		} else if (payment.manualPaymentType == true) {
+			return 'Manual Payment';
+		} else {
+			return 'Invalid Payment';
+		}
 	}
 }
